Using info instead of warning for retry_deferred and cleanup_mail (#19)
* Downgrading cleanup_mail and retry_deferred warning to info.
* Only log if there are deletions or deferreds.

+++ django_yubin/management/commands/cleanup_mail.py
@@ -31,6 +31,7 @@ class Command(BaseCommand):
         today = datetime.date.today()
         cutoff_date = today - datetime.timedelta(options['days'])
         count = Message.objects.filter(date_created__lt=cutoff_date).count()
-        Message.objects.filter(date_created__lt=cutoff_date).delete()
-        logger.warning("Deleted %s mails created before %s " %
-                       (count, cutoff_date))
+        if count:
+            Message.objects.filter(date_created__lt=cutoff_date).delete()
+            logger.info("Deleted %s mails created before %s " %
+                        (count, cutoff_date))
